Now using FEBio's vector and matrix classes for febcode.

+++ febcode/compiler.cpp
@@ -76,7 +76,7 @@ void Compiler::endScope()
 		}
 
 		m_locals.pop_back();
-		localStackSize -= local.type->size();
+		localStackSize -= (int)local.type->size();
 	}
 	m_scopeDepth--;
 }
